More detailed source attribution
This update enables extra attribution sources to be shown in the view for individual locations. It lists the source name, description, and contact info of the original as it did before.  It just adds the source names  and url in the attributions field of the api response in order to list other sources if there are others.  Tweaked app/assets/scripts/actions/measurements to get the attribution info in the api respsone and use it in app/assets/scripts/views/location. Also tweaked the jsx and app/assets/styles/location/_content to show the source names in one column and keep the description and contact info in another as well as add 'Sources:' as a section header.